requirements.txt
gensim==4.2.0
numpy>1.18.0
fastapi==0.74.1
pandas==1.4.1
uvicorn==0.17.5
python-decouple==3.8